Move parser structure to separate file
